backend/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-environ==0.4.5
djangorestframework-camel-case==1.1.2
djangorestframework-simplejwt==4.4.0
cryptography==2.8
simplejwt-extensions==0.1.0
